Create 100doors.cpp
You have 100 doors from 1 to 100 all are in a closed state initially. There are 100 children indexed from 1 to 100. The 1st student go and change the state of the doors that are multiple of 1 similarly second student go and change the state of doors that are multiple of 2 and so on. After all the 100 children print all the doors that are in closed state and all the doors in open state